+++ src/gen_cranberry/gc_dev_spanel.h
@@ -4,5 +4,5 @@
 #ifndef GC_DEV_SPANEL
 #define GC_DEV_SPANEL
 void gc_dev_spanel_open(void);
-int gc_dev_spanel_read(void);
+uint16_t gc_dev_spanel_read(void);
 #endif
